Modified on 25-08-2024

+++ src/MATHEMATICALS/NewtonRapsonMethodForSquareRoot.java
@@ -4,18 +4,18 @@ public class NewtonRapsonMethodForSquareRoot {
     public static void main(String[] args) {
         System.out.println(Square(40));
     }
-    static  double Square(double n){
+    static  int Square(double n){
         double x = n;
         double root;
 
         while (true){
 
             root = 0.5 * (x + (n/x));
-            if(Math.abs(root -x) < 0.5)
+            if(Math.abs(root - x) < 0.5)
                 break;
 
             x = root;
         }
-        return  root;
+        return (int) root;
     }
 }
